feat: sort hierarchy in cache package
