+++ packages/cache/src/key-cache/index.ts
@@ -3,27 +3,41 @@ import { Atom } from "@rixio/atom"
 import { SimpleCache } from "@rixio/lens"
 import { Subject } from "rxjs"
 import { filter, first } from "rxjs/operators"
-import { CacheImpl } from "./cache"
-import { BatchHelper } from "./key-batch"
-import * as Domain from "./domain"
-import { byKeyWithDefaultFactory } from "./utils"
+import { BatchHelper } from "../utils/batch-helper"
+import { byKeyWithDefaultFactory } from "../utils"
+import { CacheImpl } from "../cache"
+import { createAddEvent, isNotFound, KeyEvent, ListDataLoader, NotFound, UNDEFINED } from "../common/domain"
+import { CacheState, createFulfilledCache, idleCache } from "../cache/domain"
+import { Cache } from "../cache"
 
-export class KeyCacheImpl<K, V> implements Domain.KeyCache<K, V> {
+export interface KeyCache<K, V> {
+	get(key: K, force?: boolean): Promise<V>
+	set(key: K, value: V): void
+	getAtom(key: K): Atom<CacheState<V>>
+	single(key: K): Cache<V>
+}
+
+/**
+ * @deprecated this type of cache deprecated
+ * please use key-memo instead
+ */
+
+export class KeyCacheImpl<K, V> implements KeyCache<K, V> {
 	private readonly batchHelper: BatchHelper<K>
-	private readonly results = new Subject<[K, V | Domain.NotFound]>()
-	private readonly lensFactory = byKeyWithDefaultFactory<K, Domain.CacheState<V>>(Domain.idleCache)
-	private readonly _events = new Subject<Domain.KeyEvent<K>>()
+	private readonly results = new Subject<[K, V | NotFound]>()
+	private readonly lensFactory = byKeyWithDefaultFactory<K, CacheState<V>>(idleCache)
+	private readonly _events = new Subject<KeyEvent<K>>()
 	readonly events = this._events.pipe()
 
-	private readonly singles = new SimpleCache<K, Domain.Cache<V>>(key => {
+	private readonly singles = new SimpleCache<K, Cache<V>>(key => {
 		return new CacheImpl(this.getAtom(key), () => {
 			return this.load(key)
 		})
 	})
 
 	constructor(
-		private readonly map: Atom<IM<K, Domain.CacheState<V>>>,
-		private readonly loader: Domain.ListDataLoader<K, V>,
+		private readonly map: Atom<IM<K, CacheState<V>>>,
+		private readonly loader: ListDataLoader<K, V>,
 		timeout: number = 100
 	) {
 		this.onBatchLoad = this.onBatchLoad.bind(this)
@@ -39,12 +53,12 @@ export class KeyCacheImpl<K, V> implements Domain.KeyCache<K, V> {
 		} catch (e) {
 			console.error(e)
 			keys.forEach(k => {
-				this.results.next([k, Domain.UNDEFINED])
+				this.results.next([k, UNDEFINED])
 			})
 		}
 	}
 
-	single(key: K): Domain.Cache<V> {
+	single(key: K): Cache<V> {
 		return this.singles.getOrCreate(key, () => {
 			this.onCreate(key)
 		})
@@ -56,16 +70,16 @@ export class KeyCacheImpl<K, V> implements Domain.KeyCache<K, V> {
 
 	set(key: K, value: V): void {
 		this.map.modify(x => {
-			return x.set(key, Domain.createFulfilledCache(value))
+			return x.set(key, createFulfilledCache(value))
 		})
 		this.onCreate(key)
 	}
 
 	onCreate(key: K) {
-		this._events.next(Domain.createAddEvent(key))
+		this._events.next(createAddEvent(key))
 	}
 
-	getAtom(key: K): Atom<Domain.CacheState<V>> {
+	getAtom(key: K): Atom<CacheState<V>> {
 		return this.map.lens(this.lensFactory(key))
 	}
 
@@ -77,7 +91,7 @@ export class KeyCacheImpl<K, V> implements Domain.KeyCache<K, V> {
 				first()
 			)
 			.toPromise()
-		if (Domain.isNotFound(v)) {
+		if (isNotFound(v)) {
 			throw new Error(`Entity with key "${key}" not found`)
 		}
 		return v
