+++ packages/cache/src/key-memo/index.ts
@@ -3,27 +3,35 @@ import { Atom } from "@rixio/atom"
 import { SimpleCache } from "@rixio/lens"
 import { Subject } from "rxjs"
 import { filter, first } from "rxjs/operators"
-import { MemoImpl } from "./memo"
-import { BatchHelper } from "./key-batch"
-import * as Domain from "./domain"
-import { byKeyWithDefaultFactory } from "./utils"
+import { Memo, MemoImpl } from "../memo"
+import { BatchHelper } from "../utils/batch-helper"
+import { byKeyWithDefaultFactory } from "../utils"
+import { KeyEvent, NotFound, ListDataLoader, UNDEFINED, createAddEvent, isNotFound } from "../common/domain"
+import { CacheState, idleCache, createFulfilledCache } from "../cache/domain"
 
-export class KeyMemoImpl<K, V> implements Domain.KeyMemo<K, V> {
+export interface KeyMemo<K, V> {
+	get(key: K, force?: boolean): Promise<V>
+	set(key: K, value: V): void
+	getAtom(key: K): Atom<CacheState<V>>
+	single(key: K): Memo<V>
+}
+
+export class KeyMemoImpl<K, V> implements KeyMemo<K, V> {
 	private readonly batchHelper: BatchHelper<K>
-	private readonly results = new Subject<[K, V | Domain.NotFound]>()
-	private readonly lensFactory = byKeyWithDefaultFactory<K, Domain.CacheState<V>>(Domain.idleCache)
-	private readonly _events = new Subject<Domain.KeyEvent<K>>()
+	private readonly results = new Subject<[K, V | NotFound]>()
+	private readonly lensFactory = byKeyWithDefaultFactory<K, CacheState<V>>(idleCache)
+	private readonly _events = new Subject<KeyEvent<K>>()
 	readonly events = this._events.pipe()
 
-	private readonly singles = new SimpleCache<K, Domain.Memo<V>>(key => {
+	private readonly singles = new SimpleCache<K, Memo<V>>(key => {
 		return new MemoImpl(this.getAtom(key), () => {
 			return this.load(key)
 		})
 	})
 
 	constructor(
-		private readonly map: Atom<IM<K, Domain.CacheState<V>>>,
-		private readonly loader: Domain.ListDataLoader<K, V>,
+		private readonly map: Atom<IM<K, CacheState<V>>>,
+		private readonly loader: ListDataLoader<K, V>,
 		timeout: number = 100
 	) {
 		this.onBatchLoad = this.onBatchLoad.bind(this)
@@ -39,12 +47,12 @@ export class KeyMemoImpl<K, V> implements Domain.KeyMemo<K, V> {
 		} catch (e) {
 			console.error(e)
 			keys.forEach(k => {
-				this.results.next([k, Domain.UNDEFINED])
+				this.results.next([k, UNDEFINED])
 			})
 		}
 	}
 
-	single(key: K): Domain.Memo<V> {
+	single(key: K): Memo<V> {
 		return this.singles.getOrCreate(key, () => {
 			this.onCreate(key)
 		})
@@ -56,16 +64,16 @@ export class KeyMemoImpl<K, V> implements Domain.KeyMemo<K, V> {
 
 	set(key: K, value: V): void {
 		this.map.modify(x => {
-			return x.set(key, Domain.createFulfilledCache(value))
+			return x.set(key, createFulfilledCache(value))
 		})
 		this.onCreate(key)
 	}
 
 	onCreate(key: K) {
-		this._events.next(Domain.createAddEvent(key))
+		this._events.next(createAddEvent(key))
 	}
 
-	getAtom(key: K): Atom<Domain.CacheState<V>> {
+	getAtom(key: K): Atom<CacheState<V>> {
 		return this.map.lens(this.lensFactory(key))
 	}
 
@@ -77,7 +85,7 @@ export class KeyMemoImpl<K, V> implements Domain.KeyMemo<K, V> {
 				first()
 			)
 			.toPromise()
-		if (Domain.isNotFound(v)) {
+		if (isNotFound(v)) {
 			throw new Error(`Entity with key "${key}" not found`)
 		}
 		return v
